% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbsubset.R
\name{bbsubset.G}
\alias{bbsubset.G}
\title{Extract well-balanced subset minimizing abs(error) with LP performed by gurobi}
\usage{
bbsubset.G(S, k)
}
\arguments{
\item{S}{barcode set}

\item{k}{the size of barcode subset to extract}
}
\description{
Extract well-balanced subset minimizing abs(error) with LP performed by gurobi
}
