% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbsubset.R
\name{bbsubset}
\alias{bbsubset}
\title{Extract well-balanced subset minimizing abs(error) with LP}
\usage{
bbsubset(S, k, ...)
}
\arguments{
\item{S}{barcode set}

\item{k}{the size of barcode subset to extract}

\item{...}{pass to \code{ROI_solve}}
}
\description{
Extract well-balanced subset minimizing abs(error) with LP
}
