% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbsubset.R
\docType{data}
\name{sample_barcodes}
\alias{sample_barcodes}
\title{Sample barcodes}
\format{
A vactor with 76 elements
}
\usage{
sample_barcodes
}
\description{
A dataset containing 76 DNA barcode.Their length are 6 base pair and
they are three humming distances apart from each other.
This data set is created by \code{DNABarcodes}
}
\keyword{datasets}
