% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbsubset.R
\name{bbsubset}
\alias{bbsubset}
\title{min abs(error) with LP (fastest if using commercial solvers)}
\usage{
bbsubset(S, k)
}
\arguments{
\item{S}{barcode set}

\item{k}{the size of barcode subset to extract}
}
\description{
min abs(error) with LP (fastest if using commercial solvers)
}
