% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbsubset.R
\name{bbsubset}
\alias{bbsubset}
\title{Extract well-balanced subset minimizing abs(error) with LP}
\usage{
bbsubset(S, k, ...)
}
\arguments{
\item{S}{barcode set}

\item{k}{the size of barcode subset to extract}

\item{...}{pass to \code{ROI_solve}. Please see the examples below.}
}
\description{
Extract well-balanced subset minimizing abs(error) with LP
}
\examples{
\dontrun{
# a

bbsubset(sample_barcodes,k=8)

# Set a timeout

bbsubset(sample_barcodes,k=5,timeout=10)

# Use the other LP solver

bbsubset(sample_barcodes,k=8,solver="gurobi")
}

}
